requirements.txt
asn1crypto==0.24.0
azure-common==1.1.21
azure-storage-blob==2.0.1
azure-storage-common==2.0.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.6.1
Flask==1.0.3
git+https://github.com/mitsuhiko/flask-oauth   #Flask-OAuth==0.13
httplib2==0.12.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
MarkupSafe==1.1.1
oauth2==1.9.0.post1
pycparser==2.19
python-dateutil==2.8.0
requests==2.22.0
six==1.12.0
soupsieve==1.9.1
SQLAlchemy==1.3.3
urllib3==1.25.3
Werkzeug==0.15.4
